requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
multidict==6.0.4
openai==0.28.1
PyAudio==0.2.14
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SpeechRecognition==3.10.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
